initial commit of test_mount_status()
Currently incomplete but merging as-is to avoid duplication of effort.
With indicated (via added comments) tested code changes in osi.py this
test's mocking of builin open is workable but is currently awaiting the
Python 3 move where it is anticipated that the current mocking approach
will work without tested code changes.