establish new rpm/repo/project signing key ready for relaunch #2176
Our Web-UI integrated update mechanism uses the locally installed
instance of this file to authenticate the rockstor rpm updates and our
upstream zypper subsystem defaults to expecting both rpm and rpm
repositories to be signed similarly.

As this file, in it's GitHub instance, is also referenced from our forum
as the source of truth with regard to our current public key it is
required that we, shortly after merging, publish a new rpm appropriately
signed and in turn re-sign it's associated rpm repository.

Existing clients, currently only 'Built on openSUSE' alpha/beta testers,
will then have to perform a one off command line rpm import of this
file, from it's GitHub instance, to facilitate further updates from our
testing rpm repository. As we are yet to populate/establish our
'Built on openSUSE' Stable Channel updates repository it will not be
affected, and upon later initialization of that repo this key will be in
effect from the start.