Improve error message on missing device removal failure #2099
In some known cases a misleading/contradictory error message results
when attempting to remove a missing device: that of there being no
missing devices when btrfs fi show indicates otherwise. Handle this case
by re-asserting the degraded,rw mount options suggested in our
Maintenance required contextual text and further suggest a reboot may be
required. It is suggested that aim to provide an unmount/remount
capability in the future that could also be suggested here.