add recognition of system disk & pool when on SD or MicroSD card #2192
For installs using an SD or MicroSD card as the system disk there is a
failure to recognise this device as hosting the system with a consequent
failure to surface the system pool. Further to this the user is
presented with inappropriate choices; as if this disk was a regular data
device.

Fix by adding the necessary device name match, and appropriate base name
extraction filter, too root_disk(). In this case our existing mdraid
device name treatment shares the same name structure assumptions with
regard to base device name ("mmcblk1") in relation to partition name
("mmcblk1p3"). And so this patch proposes that an 'or' match is added to
this existing section.