hide non functional auto update button until post re-launch fix #2196
Pending the transition to a yast2-online-update-configuration
integration, compatible with our current openSUSE target distro, remark
out the Web-UI element to our legacy yum-cron based system. This enables
our existing related structures to serve as a guide for the transition
and removes a non functional element that will ultimately only
contribute to confusion and frustration if left in it's current
visible and non functional form.