repeated log errors on disabled quotas (#2236)
Remove explicit logging of key 'btrfs qgroup show' commands to
avoid excessive log spamming. As we have now supported the pool quota
disabled state mostly successfully for some time we can remove the
initial default of always logging these commands which are reported as
errors by the underlying btrfs commands when quotas are disabled on the
parent pool. And hence are at odds with their supported state within
the Rockstor Web-UI. In most common cases we catch and log as info in
addition to the prior 'raw' command log.

Prior explicit log parameters have been kept as explicit to ease
reversing this change.