Fake Product UUID Case Sensitivity Incorrect on Rockstor 4 (#2290)
As Rockstor 4 is associated with newer kernel versions by default,
which in turn now produce lowercase product uuid's, we fail in the
majority of cases to match against our mixed case tuple of
know fake product uuids.

Fix by normalising on lowercase in line with the sited uuid
standard of: rfc4122 as per the kernel.

Includes formatting clean-ups of the known fake product uuids.

N.B. no functional change is made to how an existing install, of the
same major version, generates it's Appliance ID.