Re-enable IPv6 at kernel level #2309
Adds systemd rockstor-ipv6check.service file to conf directory, making
it available for the "%post" section of our rpm to install, enable, and
unlike our other service, start. We then have an rpm update mechanism
for existing installs to receive this re-configuration. A consequent
reboot is then required but we are still in the testing channel so this
is acceptable. But the late release candidate nature of our current
testing channel means we need at least this level of automated re-config
to account for our less technical testers. Source installs are not
addressed as they are deemed for development only where the system's
setup is the responsibility of the developer. Our installer has this fix
pre-installed which would negate this secondary install/update mechanism
via the included conditional: i.e. this service will only execute if the
running kernel has the "ipv6.disable=1" option active.

Service summary:
- Conditionally run only if the kernel command line contains
"ipv6.disable=1"
- Remove via stream edit "ipv6.disable=1" with or without trailing space
from /etc/default/grub
- Invoke a grub re-config to instantiate the new setup.
- Echo to console and regular log the following message:
" Reboot required to re-enable IPv6 & Rock-ons " with a magenta
background effective in a standard console.