Move systemd setup from rockstor-bootstrap rockstor-pre #2432
We already to the vast majority of our initial systemd
service config in initrock.py (rockstor-pre.service), so
move the remaining duplicate firewalld and unique atd systemd
commands from the 'bootstrap' command in command.py of
storageadmin invoked by our scripts/bootstrap which in turn
is started by our rockstor-bootstrap; to our rockstor-pre
initrock.py setup. This also helps with transparency regarding
our initial systemd requirements. And keeps all these initial
settings in one place.
Also fixed rockstor-bootstrap.service now inadvertently disabling
and stopping nginx as we no longer start this with supervisord
and so we are no longer required to move the default os systemd
service 'out-of-the-way' as we now use it and override it's config.