include files sufficient for our sdist to self build #2432
We are currently aiming at a poetry/pyproject.toml maintained
and distributed source, to help keep development and deployment
managed by the same parties. As such we do not need to generate
a setup.py.
For distribution, we need to include our jslibs in the generated
sdist to avoid further dependencies during the build/deployment
process.