avoid SRPMs causing false positive update available #2458
We use the "yum changelog rockstor" command to indicate
if an update is available, and to retrieve that updates
changelog for display within the Web-UI, along with the
version of the update. When moving to a mult-arch repo
this mechanism fails as older source rpms changelogs would
be shown and confused our existing parser. Fix by using
a *.src exclude in our Rockstor-{Testing,Stable).repo yum
repo definition files.