Add form validation for share selected multiple times #2064
Selecting the same share to be bound to multiple volumes of the
same container is not yet supported by Rockstor. Add a new
validation rule isDuplicate to detect such instances of a share
being selected as the option for more than one volume and return
a specific error message accordingly.