Move Rock-on url default definitions to runtime #2305
We currently define the location of the remote and local
Rock-Ons metastores in a constant placed in settings.py.
As a result, any need to modify any of these settings would
require the user to rebuild Rockstor.

This commit moves the definition of this constant to rockon.py
where it is used, enabling the user to only need a restart of
the rockstor service after customization.