Web-UI update to start rockstor-bootstrap.service #2488
Change from starting rockstor.service to
rockstor-bootstrap.service after a Web-UI initiated
update. This service start is now in-line with that
defined by our systemd-presets-branding-rockstor
package on OBS, and with our build instructions.

We also now stop all rockstor* services
before these updates to have systemd handle the
interdependencies/ordering, and account for any
additional services of ours in the future.

## Caveat: rockstor-hdparm.service is not accounted
for: however its function is primarily during boot
and it is of Type=oneshot.