Convert pool ID in backup file to the one in the database #2508
We currently use the pool ID for a scrub scheduled task in a config
backup file as is. If this pool now has a different ID in the target
system (the one where the config backup file is being restored), this
will lead to a failure in restoring the pool scrub task.

This commit thus implements a simple conversion logic in which we fetch
the pool ID for the given pool name from the target system's database.

Update unit tests accordingly.