alt arch causing false positive update available #2468
As we are now multi arch we need to specify the arch
of the package we are interested in as the dnf-yum
changelog facility we use to retrieve this info seems
to return changelogs for all architectures.
Fix by adding "*arch" where arch = host architecture
to our "rockstor" package search term.

Add TODO to associated tests as we don't yet have this
procedure covered.