Account for empty meta of scheduled tasks of type reboot #2514
We currently assume a taskdefinition meta is always filled with information
when running the reboot_shutdown.py script. While this is the case for tasks
of type shutdown and suspend, tasks of type reboot only have an empty Dict
as meta.

This commit thus first checks for an empty meta before trying to extract
information from it.

Add unit tests coverage for reboot_shutdown.py.