improve behaviour on missing /dev/disk/by-id dir #2401
By catching the instance of no by-id directory we avoid
an inadvertent 'general' error and enable our existing
Web-UI info on missing device serial numbers with its
embedded link to our 'Minimum system requirements' docs
section. An info log is also added when this directory
is nonexistent, again with a pointer to the same doc
section.