feat(UI): adds scrollbar on large env vars forms
In case of a RockOn with many environment vars, the form will be very long. To the point that it will continue outside the viewport. I have set its max-height to 80vh, which will keep the form nice and short. But in the case of larger forms, it will keep the form within the viewport. In the latter case, a vertical scrollbar is added to allow for a nicer user experience.