use cryptsetup to retrieve container uuid #2545
Previously we used lsblk for this purpose, but
sporadically lsblk would return an empty uuid for
our just-created LUKS container. Thought to be due
to it not having yet been updated re system state.
Hence the use of the more specific LUKS command.

## Include
- New get_luks_container_uuid(disk_byid) in luks.py.
- Fresh black formatting.
- Minor comment improvements.
- Leaving in-place the prior general procedure
lsblk based mechanism as this may server elsewhere.