NetworkWidget: save data as a List object #2577
In Python2, we directly saved our processed parsing of network
activity directly from `map()`. In Python3, `map()` now returns an
iterator which prevents us from picking each of its included data
one at a time as we used to do (this now only works once).

This commit saves the `map()` object as a List so that we can
extract from it as needed.