Restore config backup functionality #2569
Following our move to Py3.6, we were unable to create a new config backup
due to a failure to get the md5 digest of a gzip file (such as a config
backup). Our config backup restore was also failing to the presence of a
Py2.7-specific bit of code.

This commit:
- adjusts `md5sum()` to be able to read a compressed file as well.
- ensures the task def validation returns the share/pool id as string
- add type hints to help ensure proper types are used/returned
- minor docstrings adjustments
- Copyright update