Catch DBusException to not throw error on LDAP group lookup #2673
We currently use InfoPipe to fetch the groupname of domain users. While
this works for Active Directory users, this can fail for LDAP users if
these do not belong to a posixAccount group.

This commit simply catches the DBusException thrown in such cases and
logs the event (DEBUG level). The result is a groupname of "null".

Also includes:
- minor Black formatting
- type hints
- import optimization
- minor copyright updates