Don't throw exception when getting supervisord service status #2681
We currently throw an Exception when checking the status of a
surpervisor-controlled service such as replication. This leads to an
exception being thrown even when expected (service is off), which can
itself interrupt other parts such as our get_services websocket.

This commit adds a new arg to the superctl() function to allow for the
possibility to not throw an exception. Note that run_command's default
for throw is True, so respect this here.