Use single https session to retrieve all rock-on definitions #2707
By sharing a single https session, instantiated on the first GET,
we avoid major cumulative delays and reduce client/server resource
overhead. Note that we still employ a dedicated https session to
retrieve our 'index' file of 'root.json'.

## Includes
- Minor fstring adoption in modified area of code.
- Info log of definitions retrieval and initial parsing time.
Intended as a diagnostic aid, and to assist with field
feedback on this process of indeterminate duration.
- Individual file retrieval timeout reduced to 5 seconds.