Use regular expressions to validate tailscale hostname #2714
The current mechanism uses a hard-coded list of characters to exclude
them from a custom tailscale hostname if present. This is rather ugly
and does not scale to all the characters (such as unicode) that need to
be excluded.

Switch to using regular expressions to exclude only characters from the
hostname that do not belong to that regular expression.