Pi4 fails on reboot #2843
Add `After` and `Wants` to rockstor-pre.service to gain
sensitivity to chrony-wait.service. This at least works
around a prior failure in starting rockstor-pre.service
on initial, and all subsequent reboots on a Pi4:
non-persistent RTC (Real Time Clock).

The very first boot was successful however: thought to be
related to its installer instance JeOS first-boot wizard
and the context of having also to wait on
rockstor-build.service (venv). Both of which modify the
service startup timing considerably.