replication received_uuid blocker re snap to share promotion #2902
When promoting the oldest of the 3 read-only snapshots received & retained
by the replication service (btrfs send/receive wrapper), use the force flag
during ro-to-rw/snap-to-share transition. At the time of this transition,
this received subvol is no longer used for comparison in all future
replication (btrfs send/receive) events. It represents an older version
of the sending systems associated replication source share. Necessarily
older by way of the constraints of the btrfs send/receive architecture,
and the safeguards of the replication wrapper: a cascade of ro snapshots.