Swap order of newVolumes representation in summary table #2933
The internal:external representation of volumes:shares in the summary
table showed at the end of the Add Storage to a Rock-On process was
the inverse to what it should have been (inverse of table headers).

This commit adjusts the Handlebars helper used to display the newly
added volume to swap the internal:external representations.